refactor: adopt UIScene lifecycle in Examples app
Add SceneDelegate to handle window ownership, URL contexts, and
WeChat callbacks. Slim AppDelegate to SDK init and appearance setup.
Replace UIMainStoryboardFile with UIApplicationSceneManifest in
Info.plist. Also fixes undefined handleSchemaURL call when
WechatOpenSDKDynamic is not available.
